move reusable functions to factory
